Make access token lifetime configurable
AB#5517
