+++ src/NHSD.BuyingCatalogue.Identity.Api/appsettings.Development.json
@@ -4,6 +4,7 @@
       "clientId": "SampleClient",
       "clientName": "Sample Client",
       "allowedGrantTypes": "Code",
+      "accessTokenLifetime": 3600,
       "allowOfflineAccess": false,
       "requireClientSecret": true,
       "secret": "SampleClientSecret",
@@ -16,6 +17,7 @@
       "clientId": "MvcClient",
       "clientName": "Mvc Client",
       "allowedGrantTypes": "Code",
+      "accessTokenLifetime": 3600,
       "allowOfflineAccess": false,
       "requireClientSecret": true,
       "secret": "MvcSecret",
@@ -28,6 +30,7 @@
       "clientId": "TokenClient",
       "clientName": "Token Client",
       "allowedGrantTypes": "ClientCredentials",
+      "accessTokenLifetime": 3600,
       "allowOfflineAccess": false,
       "requireClientSecret": true,
       "secret": "TokenSecret",
@@ -46,12 +49,5 @@
       "resourceName": "Organisation",
       "displayName": "Organisation"
     }
-  ],
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
-    }
-  }
+  ]
 }
