+++ src/NHSD.BuyingCatalogue.Identity.Api/appsettings.json
@@ -10,6 +10,7 @@
       "clientId": "SampleClient",
       "clientName": "Sample Client",
       "allowedGrantTypes": "Code",
+      "accessTokenLifetime":  3600, 
       "allowOfflineAccess": false,
       "requireClientSecret": true,
       "secret": "SampleClientSecret",
@@ -39,7 +40,7 @@
   ],
   "ConnectionStrings": {
     "CatalogueUsers": "Data Source=localhost,1451;Initial Catalog=CatalogueUsers;MultipleActiveResultSets=true;User ID=NHSD;Password=DisruptTheMarket1!"
-  },
+    },
   "cookieExpiration": {
     "expireTimeSpan": "08:00:00",
     "slidingExpiration": false
